+++ src/zopyx/typesense/interfaces.py
@@ -18,7 +18,10 @@ class IBrowserLayer(IDefaultBrowserLayer):
 class ITypesenseSettings(Interface):
     """Connector settings"""
 
-    enabled = schema.Bool(title=_("Typesense integration enabled"), default=True)
+    enabled = schema.Bool(
+        title=_("Typesense integration enabled"),
+        default=True,
+    )
 
     collection = schema.TextLine(
         title=_("Name of Typesense collection"),
@@ -27,11 +30,15 @@ class ITypesenseSettings(Interface):
     )
 
     api_key = schema.TextLine(
-        title=_("Typesense Admin API key"), default="", required=True,
+        title=_("Typesense Admin API key"),
+        default="",
+        required=True,
     )
 
     search_api_key = schema.TextLine(
-        title=_("Typesense search API key"), default="", required=True,
+        title=_("Typesense search API key"),
+        default="",
+        required=True,
     )
 
     node1_url = schema.TextLine(
@@ -66,11 +73,14 @@ class ITypesenseSettings(Interface):
     )
 
     collection_schema = schema.Text(
-        title=_("Collection schema"), default=COLLECTION_SCHEMA_JSON, required=True,
+        title=_("Collection schema"),
+        default=COLLECTION_SCHEMA_JSON,
+        required=True,
     )
 
     use_searchabletext = schema.Bool(
-        title=_("Use SearchableText for indexing as default"), default=False,
+        title=_("Use SearchableText for indexing as default"),
+        default=False,
     )