+++ CHANGES.rst
@@ -1,9 +1,15 @@
 Changelog
 =========
 
-1.0a6 (unreleased)
+1.0a6 (2022-02-19)
 ------------------
 - hide/show search results/stats based on search input state
+  [zopyx]
+- support for indexing HTML headlines (h1-h6 tags) into a dedicated
+  field
+  [zopyx]
+- highlighting support for title | headlines | text with query
+  weights 4:2:1
   [zopyx]